Started work on connecting the founder spend funds logic
